Exp: Ramsey, Relaxation; App: T1, T2
I encapsulated Ramsey and Relaxation with class methods. I also copied the codes of T1 and T2 in the Application folder and modified them to use the new exp codes. I did not import the visualization methods from QCAT yet because I only know how to do so by specifying the project directory (I think we want to install QCAT instead). All the