Modified data and plot saving method
I added a create_folder function in save_data and used it to create a folder in the experiment directory with a user specified name (very important!). If the folder already exists, the program will return a warning message along with a useless directory. S2, and S3 have been modified and tested to work with this method. I don't think S1 needs to be saved, and I plan to work on S4~S9 once I understand the structure of their data and plots.